Move routing algorithm functions into separate file
TileUtils.kt was getting a bit large, this change divides it up into
files with more specific goals.
